--- src/components/Booth.js
@@ -1,10 +1,0 @@
-import React from "react";
-import { styled } from "styled-components";
-
-const Booth = () => {
-  return <Wrapper></Wrapper>;
-};
-
-export default Booth;
-
-const Wrapper = styled.div``;
